Clonable ConfigItem,added interval in appconfig

+++ OutageManagementSystem/SCADA/SCADA_Config_Data/Repository/DataModelRepository.cs
@@ -15,7 +15,7 @@ namespace Outage.SCADA.SCADA_Config_Data.Repository
         public ushort TcpPort { get; protected set; }
         public byte UnitAddress { get; protected set; }
         public string ServiceAddress { get; protected set; }
-        public int Interval { get; protected set; }
+        public ushort Interval { get; protected set; }
 
         //public FunctionExecutor functionExecutor { get; set; }
         private INetworkModelGDAContract gdaQueryProxy = null;
@@ -46,7 +46,7 @@ namespace Outage.SCADA.SCADA_Config_Data.Repository
             TcpPort = ushort.Parse(ConfigurationManager.AppSettings["TcpPort"]);
             UnitAddress = byte.Parse(ConfigurationManager.AppSettings["UnitAddress"]);
             ServiceAddress = ConfigurationManager.AppSettings["ServiceAddress"];
-            Interval = Int32.Parse(ConfigurationManager.AppSettings["Interval"]);
+            Interval = ushort.Parse(ConfigurationManager.AppSettings["Interval"]);
             try
             {
                 gdaQueryProxy = new ChannelFactory<INetworkModelGDAContract>(EndpointNames.NetworkModelGDAEndpoint).CreateChannel();
